Internal API

PowerSimulations.add_constraints! — Method

Add power variable limits constraints for abstract hydro dispatch formulations

PowerSimulations.add_constraints! — Method

Add power variable limits constraints for abstract hydro unit commitment formulations

PowerSimulations.add_constraints! — Method

Add output power variable limits constraints for abstract hydro dispatch formulations

PowerSimulations.add_constraints! — Method

Add energy capacity down constraints for PowerSystems.HydroPumpedStorage

PowerSimulations.add_constraints! — Method

This function defines the constraints for the water level (or state of charge) for the PowerSystems.HydroPumpedStorage.

PowerSimulations.add_constraints! — Method

Add energy target constraints for PowerSystems.HydroGen

PowerSimulations.add_constraints! — Method

Time series constraints

PowerSimulations.add_constraints! — Method

Add semicontinuous range constraints for HydroCommitmentRunOfRiver formulation

PowerSimulations.add_constraints! — Method

This function defines the constraints for the water level (or state of charge) for the PowerSystems.HydroEnergyReservoir.

PowerSimulations.add_constraints! — Method

Add input power variable limits constraints for abstract hydro dispatch formulations

PowerSimulations.add_feedforward_constraints! — Method

    add_feedforward_constraints(container::PSI.OptimizationContainer,
                    model::PSI.DeviceModel,
                    devices,
                    ff::ReservoirLimitFeedforward)

Constructs a parameterized integral limit constraint to implement feedforward from other models. The Parameters are initialized using the upper boundary values of the provided variables.

sum(variable[var_name, t] for t in 1:affected_periods)/affected_periods <= param_reference[var_name]

LaTeX

$\sum_{t} x \leq param^{max}$

Arguments

container::PowerSimulations.OptimizationContainer : the optimization_container model built in PowerSimulations

model::PowerSimulations.DeviceModel : the device model

devices::IS.FlattenIteratorWrapper{T} : list of devices

ff::ReservoirLimitFeedforward : a instance of the Reservoir Limit Feedforward

PowerSimulations.add_feedforward_constraints! — Method

    add_feedforward_constraints(
        container::PSI.OptimizationContainer,
        ::PSI.DeviceModel,
        devices::IS.FlattenIteratorWrapper{T},
        ff::ReservoirTargetFeedforward,
    ) where {T <: PSY.Component}

Constructs a equality constraint to a fix a variable in one model using the variable value from other model results.

variable[var_name, t] + slack[var_name, t] >= param[var_name, t]

LaTeX

$x + slack >= param$

Arguments

container::PSI.OptimizationContainer : the optimization_container model built in PowerSimulations

model::PowerSimulations.DeviceModel : the device model

devices::IS.FlattenIteratorWrapper{T} : list of devices

ff::ReservoirTargetFeedforward : a instance of the ReservoirTarget Feedforward

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroCommitmentReservoirBudget Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroCommitmentReservoirBudget Formulation

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroCommitmentRunOfRiver Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroCommitmentRunOfRiver Formulation

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroDispatchRunOfRiver Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroDispatchRunOfRiver Formulation

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroCommitmentReservoirStorage Formulation with only Active Power

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroCommitmentReservoirStorage Formulation

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroPumpedStorage with HydroDispatchPumpedStorage Formulation with only Active Power

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroDispatchReservoirBudget Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with ReservoirBudget Dispatch Formulation

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroDispatchReservoirStorage Formulation with only Active Power

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen with HydroDispatchReservoirStorage Formulation

PowerSimulations.construct_device! — Method

Construct model for PowerSystems.HydroGen` with PowerSimulations.FixedOutput Formulation

PowerSimulations.get_min_max_limits — Method

Min and max active Power Variable limits

PowerSimulations.get_min_max_limits — Method

Min and max active Power Variable limits

PowerSimulations.get_min_max_limits — Method

Min and max input active power variable limits for HydroDispatchPumpedStorage

PowerSimulations.get_min_max_limits — Method

Min and max output active power variable limits for HydroDispatchPumpedStorage

PowerSimulations.get_min_max_limits — Method

Min and max reactive Power Variable limits