Internal API

HydroPowerSimulations.EnergyBudgetTimeSeriesParameter — Type

Parameter to define energy budget time series

HydroPowerSimulations.EnergyTargetTimeSeriesParameter — Type

Parameter to define energy storage target level time series

HydroPowerSimulations.HydroEnergySurplusVariable — Type

Struct to dispatch the creation of a slack variable for energy storage levels > target storage levels

Docs abbreviation: $E^{surplus}$

HydroPowerSimulations.HydroEnergyVariableDown — Type

Struct to dispatch the creation of a variable for energy storage level (state of charge) of lower reservoir

Docs abbreviation: $E^{down}$

HydroPowerSimulations.InflowTimeSeriesParameter — Type

Parameter to define energy inflow to storage or reservoir time series

HydroPowerSimulations.OutflowTimeSeriesParameter — Type

Parameter to define energy outflow from storage or reservoir time series

PowerSimulations.add_constraints! — Method

Add power variable limits constraints for hydro dispatch formulation

PowerSimulations.add_constraints! — Method

Add power variable limits constraints for hydro unit commitment formulation

PowerSimulations.add_constraints! — Method

Add output power variable limits constraints for hydro dispatch formulation

PowerSimulations.add_constraints! — Method

Add energy capacity down constraints for hydro pumped storage

PowerSimulations.add_constraints! — Method

This function defines the constraints for the water level (or state of charge) for the HydroPumpedStorage.

PowerSimulations.add_constraints! — Method

Add energy target constraints for hydro gen

PowerSimulations.add_constraints! — Method

Time series constraints

PowerSimulations.add_constraints! — Method

Add semicontinuous range constraints for Hydro Unit Commitment formulation

PowerSimulations.add_constraints! — Method

This function defines the constraints for the water level (or state of charge) for the Hydro Reservoir.

PowerSimulations.add_constraints! — Method

Add input power variable limits constraints for hydro dispatch formulation

PowerSimulations.add_feedforward_constraints! — Method

    add_feedforward_constraints(container::OptimizationContainer,
                    cons_name::Symbol,
                    param_reference,
                    var_key::VariableKey)

Constructs a parameterized integral limit constraint to implement feedforward from other models. The Parameters are initialized using the upper boundary values of the provided variables.

sum(variable[var_name, t] for t in 1:affected_periods)/affected_periods <= param_reference[var_name]

LaTeX

$\sum_{t} x \leq param^{max}$

Arguments

container::OptimizationContainer : the optimization_container model built in PowerSimulations

model::DeviceModel : the device model

devices::IS.FlattenIteratorWrapper{T} : list of devices

ff::FixValueFeedforward : a instance of the FixValue Feedforward

PowerSimulations.add_feedforward_constraints! — Method

    add_feedforward_constraints(
        container::PSI.OptimizationContainer,
        ::PSI.DeviceModel,
        devices::IS.FlattenIteratorWrapper{T},
        ff::ReservoirTargetFeedforward,
    ) where {T <: PSY.Component}

Constructs a equality constraint to a fix a variable in one model using the variable value from other model results.

variable[var_name, t] + slack[var_name, t] >= param[var_name, t]

LaTeX

$x + slack >= param$

Arguments

container::PSI.OptimizationContainer : the optimization_container model built in PowerSimulations

model::PSI.DeviceModel : the device model

devices::IS.FlattenIteratorWrapper{T} : list of devices

ff::ReservoirTargetFeedforward : a instance of the FixValue Feedforward

PowerSimulations.construct_device! — Method

Construct model for HydroGen with ReservoirBudget Commitment Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for HydroGen with ReservoirBudget Commitment Formulation

PowerSimulations.construct_device! — Method

Construct model for HydroGen with RunOfRiver Commitment Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for HydroGen with RunOfRiver Dispatch Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for HydroGen with RunOfRiver Dispatch Formulation

PowerSimulations.construct_device! — Method

Construct model for HydroGen with ReservoirStorage Dispatch Formulation with only Active Power

PowerSimulations.construct_device! — Method

Construct model for HydroGen with ReservoirStorage Commitment Formulation

PowerSimulations.construct_device! — Method

Construct model for HydroPumpedStorage with PumpedStorage Dispatch Formulation with only Active Power

PowerSimulations.construct_device! — Method

Construct model for HydroGen with ReservoirBudget Dispatch Formulation with only Active Power.

PowerSimulations.construct_device! — Method

Construct model for HydroGen with ReservoirBudget Dispatch Formulation

PowerSimulations.construct_device! — Method

Construct model for HydroGen with ReservoirStorage Dispatch Formulation with only Active Power

PowerSimulations.construct_device! — Method

Construct model for HydroGen with ReservoirStorage Dispatch Formulation

PowerSimulations.construct_device! — Method

Construct model for HydroGen with FixedOutput Formulation

PowerSimulations.get_min_max_limits — Method

Min and max active Power Variable limits

PowerSimulations.get_min_max_limits — Method

Min and max active Power Variable limits

PowerSimulations.get_min_max_limits — Method

Min and max input active power variable limits for hydro dispatch pumped storage

PowerSimulations.get_min_max_limits — Method

Min and max output active power variable limits for hydro dispatch pumped storage

PowerSimulations.get_min_max_limits — Method

Min and max reactive Power Variable limits