Public API Reference

HydroPowerSimulations.HydroCommitmentReservoirBudget — Type

Formulation type to add commitment and injection variables constrained by total energy production budget defined with a time series for HydroGen

HydroPowerSimulations.HydroCommitmentReservoirStorage — Type

Formulation type to constrain hydropower production with unit commitment variables and a representation of the energy storage capacity and water inflow time series of a reservoir for HydroGen

HydroPowerSimulations.HydroCommitmentRunOfRiver — Type

Formulation type to add commitment and injection variables constrained by a maximum injection time series for HydroGen

HydroPowerSimulations.HydroDispatchPumpedStorage — Type

Formulation type to constrain energy production from pumped storage with a representation of the energy storage capacity of upper and lower reservoirs and water inflow time series of upper reservoir and outflow time series of lower reservoir for HydroPumpedStorage

HydroPowerSimulations.HydroDispatchReservoirBudget — Type

Formulation type to add injection variables constrained by total energy production budget defined with a time series for HydroGen

HydroPowerSimulations.HydroDispatchReservoirStorage — Type

Formulation type to constrain hydropower production with a representation of the energy storage capacity and water inflow time series of a reservoir for HydroGen

HydroPowerSimulations.HydroDispatchRunOfRiver — Type

Formulation type to add injection variables constrained by a maximum injection time series for HydroGen

HydroPowerSimulations.HydroEnergyOutput — Type

Auxiliary Variable for Hydro Models that solve for total energy output

HydroPowerSimulations.HydroEnergyShortageVariable — Type

Struct to dispatch the creation of a slack variable for energy storage levels < target storage levels

Docs abbreviation: $E^{shortage}$

HydroPowerSimulations.HydroEnergyVariableUp — Type

Struct to dispatch the creation of a variable for energy storage level (state of charge) of upper reservoir

Docs abbreviation: $E^{up}$

HydroPowerSimulations.ReservoirLimitFeedforward — Type

Adds a constraint to limit the sum of a variable over the number of periods to the source value

HydroPowerSimulations.ReservoirLimitParameter — Type

Parameter to define energy limit

HydroPowerSimulations.ReservoirTargetFeedforward — Type

Adds a constraint to enforce a minimum reservoir level target with a slack variable associated with a penalty term.

Fields:

optmization_container_key ->
source - From where the data comes
affected_values -

HydroPowerSimulations.ReservoirTargetParameter — Type

Parameter to define energy target

HydroPowerSimulations.WaterSpillageVariable — Type

Struct to dispatch the creation of energy (water) spillage variable representing energy released from a storage/reservoir not injected into the network

Docs abbreviation: $S$